@@ -1,11 +1,15 @@
 module github.com/biogo/hts
 
-go 1.14
+go 1.19
 
 require (
 	github.com/biogo/boom v0.0.0-20150317015657-28119bc1ffc1
 	github.com/kortschak/utter v0.0.0-20190412033250-50fe362e6560
-	github.com/kr/pretty v0.2.0 // indirect
 	github.com/ulikunitz/xz v0.5.10
 	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15
+)
+
+require (
+	github.com/kr/pretty v0.2.0 // indirect
+	github.com/kr/text v0.1.0 // indirect
 )
